+++ frontend-app/src/app/layouts/admin-layout/admin-layout.routing.ts
@@ -55,6 +55,7 @@ export const AdminLayoutRoutes: Routes = [
     // }
     { path: 'dashboard',      component: DashboardComponent },
     { path: 'user-profile',   component: UserProfileComponent },
+    { path: 'user-profile/:id', component: UserProfileComponent },
     { path: 'user-list',      component: TableListComponent },
     { path: 'typography',     component: TypographyComponent },
     { path: 'icons',          component: IconsComponent },